\name{alignedPeakList}
\alias{alignedPeakList}
\docType{data}
\title{Example Data}
\description{
List containing information of time labels for peaks and peak intensity value for corresponding spectrum
}
\usage{data(alignedPeakList)}
\format{The format is List of 2}
\examples{
data(alignedPeakList)
}
\keyword{datasets}
