\name{miCalc}
\alias{miCalc}
\title{C Code to Speed the Calculation of Mutual Information}
\description{C Code to Speed the Calculation of Mutual Information}
\usage{
miCalc(v1, v2, n, MI)
}
\arguments{
  \item{v1}{one of the two vectors}
  \item{v2}{one of the two vectors}
  \item{n}{length of the vector}
  \item{MI}{mutual information of the two vectors}
}
\value{mutual information of the two vectors}
\author{Bill Cooke, College of William and Mary, Dept. of Physics, 2009}
