\name{repeats}
\alias{repeats}
\docType{data}
\title{Example Data}
\description{Integer, times to repeat the whole process (e.g.re-crossvalidate). 2 is used here for the example data, but 100 is recommended when processing the real data}
\usage{data(repeats)}
\examples{
data(repeats)
}
\keyword{datasets}
