\name{rWMBAT-package}
\alias{rWMBAT-package}
\alias{rWMBAT}
\docType{package}
\title{The William and Mary Bayesian Analysis Tool}
\description{The package contains all the functions used to get a resulting Bayesian network. Then main function is "WMBAT", which takes an array of mass spec peak intensities and other information to build and assess a Bayesian network, the primary output is an adjacency matrix describing the resulting Bayesian network. Other functions, e.g., DoTheMath, are sub-functions used in the "WMBAT" to get the final result.}
\details{
\tabular{ll}{
Package: \tab rWMBAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Karl Kuschner, Qian Si and William Cooke, College of William and Mary, Dept. of Physics, 2009}
\references{http://kwkusc.people.wm.edu/dissertation/dissertation.htm}
