\name{threshold}
\alias{threshold}
\docType{data}
\title{Example Data}
\description{
Factor by which the maximum "random" MI is multiplied to find the minimum "significant" MI (double, 1.0-5.0)
}
\usage{data(threshold)}
\details{
recommend starting with 1 and increasing until a "reasonable"
number of diagnostic peaks is reached and error rates are
minimized. This setting is dependant on the data and the
correlations between variables.
}
\keyword{datasets}
