\name{tofListMetaData}
\alias{tofListMetaData}
\docType{data}
\title{Example Data}
\description{One of the input example data for main function "WMBAT", a list containing information about every spectra, the information related to the spectra we needed in this package is class and ID.
Class,Integer vector, values 1 or 2 identifying the class of each case, such as "disease, non-disease"
ID,Double one or two column array contains the sample IDfor each case. Second column is optional and would identifyreplicates of the same sample.  
}
\usage{data(tofListMetaData)}
\keyword{datasets}
