\name{rWMBAT-package}
\alias{rWMBAT-package}
\alias{rWMBAT}
\docType{package}
\title{The William and Mary Bayesian Analysis Tool}
\description{builds and assesses a Bayesian network after selecting features (peaks) from within the data array that are diagnostic of the class}
\details{
\tabular{ll}{
Package: \tab rWMBAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-04\cr
License: \tab GPL(version 2 or later)\cr
LazyLoad: \tab yes\cr
}
}
\author{Karl Kuschner, Qian Si and William Cooke, College of William and Mary, Dept. of Physics, 2009}
\references{http://kwkusc.people.wm.edu/dissertation/dissertation.htm}
